requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
kafka-python==2.0.2
